requirements.txt
opencv-contrib-python-headless
Flask==2.0.2
flask_cors==3.0.10
matplotlib==3.3.3
numpy==1.19.5
pandas==1.1.5
tensorflow-cpu
joblib